participant-template/credentials-manager/api/app/requirements.txt
fastapi
pydantic
uvicorn
fastapi_keycloak
PyJWT==2.1.0
python-multipart
rich==13.7.0
requests==2.31.0
environ-config==23.2.0
sh==2.0.6
cryptography==42.0.5
jwcrypto==1.5.4
coloredlogs==15.0.1
sqlalchemy==2.0.30
pyld==2.0.4
jcs==0.2.1